NIFI-8273 Cleaning up and adding additional details based on review comment
