+++ nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/main/java/org/apache/nifi/processors/script/ScriptedPartitionRecord.java
@@ -76,7 +76,6 @@ import java.util.Set;
 })
 @SeeAlso(classNames = {
         "org.apache.nifi.processors.script.ScriptedTransformRecord",
-        "org.apache.nifi.processors.script.ScriptedRouteRecord",
         "org.apache.nifi.processors.script.ScriptedValidateRecord",
         "org.apache.nifi.processors.script.ScriptedFilterRecord"
 })
@@ -150,9 +149,12 @@ public class ScriptedPartitionRecord extends ScriptedRecordProcessor {
             final FlowFile incomingFlowFile,
             final ScriptEvaluator evaluator
     ) {
+        final long startMillis = System.currentTimeMillis();
+
         final RecordReaderFactory readerFactory = context.getProperty(RECORD_READER).asControllerService(RecordReaderFactory.class);
         final RecordSetWriterFactory writerFactory = context.getProperty(RECORD_WRITER).asControllerService(RecordSetWriterFactory.class);
         final Map<String, String> originalAttributes = incomingFlowFile.getAttributes();
+        final RecordCounts counts = new RecordCounts();
 
         try {
             session.read(incomingFlowFile, new InputStreamCallback() {
@@ -168,13 +170,12 @@ public class ScriptedPartitionRecord extends ScriptedRecordProcessor {
                         final Map<String, FlowFile> outgoingFlowFiles = new HashMap<>();
                         final Map<String, RecordSetWriter> recordSetWriters = new HashMap<>();
 
-                        int index = 0;
-
                         // Reading in records and evaluate script
                         while (pushBackSet.isAnotherRecord()) {
                             final Record record = pushBackSet.next();
-                            final Object evaluatedValue = evaluator.evaluate(record, index++);
-                            getLogger().debug("Evaluated scripted against {} (index {}), producing result of {}", record, index - 1, evaluatedValue);
+                            final Object evaluatedValue = evaluator.evaluate(record, counts.getRecordCount());
+                            getLogger().debug("Evaluated scripted against {} (index {}), producing result of {}", record, counts.getRecordCount(), evaluatedValue);
+                            counts.incrementRecordCount();
 
                             final String partition = (evaluatedValue == null) ? null : evaluatedValue.toString();
                             RecordSetWriter writer = recordSetWriters.get(partition);
@@ -219,16 +220,19 @@ public class ScriptedPartitionRecord extends ScriptedRecordProcessor {
                             fragmentIndex++;
                         }
 
-                        session.adjustCounter("Record Processed", index, false);
+                        final long millis = System.currentTimeMillis() - startMillis;
+                        session.adjustCounter("Records Processed", counts.getRecordCount(), true);
+                        session.getProvenanceReporter().fork(incomingFlowFile, outgoingFlowFiles.values(), "Processed " + counts.getRecordCount() + " Records", millis);
+
                     } catch (final ScriptException | SchemaNotFoundException | MalformedRecordException e) {
-                        throw new ProcessException("Failed to parse incoming FlowFile", e);
+                        throw new ProcessException("After processing " + counts.getRecordCount() +  " Records, encountered failure when attempting to process " + incomingFlowFile, e);
                     }
                 }
             });
 
             return true;
         } catch (final Exception e) {
-            getLogger().error("Failed to route records for {}", incomingFlowFile, e);
+            getLogger().error("Failed to partition records due to: " + e.getMessage(), e);
             return false;
         }
     }
