+++ nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/main/java/org/apache/nifi/processors/script/ScriptedValidateRecord.java
@@ -35,7 +35,6 @@ import java.util.Set;
 @SeeAlso(classNames = {
         "org.apache.nifi.processors.script.ScriptedTransformRecord",
         "org.apache.nifi.processors.script.ScriptedFilterRecord",
-        "org.apache.nifi.processors.script.ScriptedRouteRecord",
         "org.apache.nifi.processors.script.ScriptedPartitionRecord"
 })
 public class ScriptedValidateRecord extends ScriptedRouterProcessor<Boolean> {
