--- nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/main/resources/docs/org.apache.nifi.processors.script.ScriptedRouteRecord/additionalDetails.html
@@ -1,104 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns="http://www.w3.org/1999/html">
-<!--
-      Licensed to the Apache Software Foundation (ASF) under one or more
-      contributor license agreements.  See the NOTICE file distributed with
-      this work for additional information regarding copyright ownership.
-      The ASF licenses this file to You under the Apache License, Version 2.0
-      (the "License"); you may not use this file except in compliance with
-      the License.  You may obtain a copy of the License at
-          http://www.apache.org/licenses/LICENSE-2.0
-      Unless required by applicable law or agreed to in writing, software
-      distributed under the License is distributed on an "AS IS" BASIS,
-      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-      See the License for the specific language governing permissions and
-      limitations under the License.
-    -->
-
-<head>
-    <meta charset="utf-8"/>
-    <title>ScriptedRouteRecord</title>
-    <link rel="stylesheet" href="../../../../../css/component-usage.css" type="text/css"/>
-    <style>
-h2 {margin-top: 4em}
-h3 {margin-top: 3em}
-td {text-align: left}
-    </style>
-</head>
-
-<body>
-
-<h1>ScriptedRouteRecord</h1>
-
-<h3>Description</h3>
-
-<p>
-    The ScriptedRouteRecord provides the ability to use a scripting language, such as Groovy or Jython, to quickly and easily route a Record based on the contents of it.
-    There are multiple ways to reach the same behaviour such as using PartitionRecord and RouteOnAttribute but besides the performance cost of the multiple processing steps,
-    working with user provided scripts opens a wide range of possibilities on the decision logic of routing the individual records.
-</p>
-
-<h3>Script and routes</h3>
-
-<p>
-    The provided script is evaluated once for each Record that is encountered in the incoming FlowFile. Each time that the script is invoked, it is expected to return a string value.
-    The return value is used as "route". A "route" is considered as a batch of Records intended to be managed together. As of, Records in a route will be added to one outgoing FlowFile.
-</p>
-
-<h3>Relationships</h3>
-
-<p>
-    Every route is bound to an outgoing Relationship. The FlowFile related to the given route will will be transferred to the bound Relationship. The bounding between a route and a
-    Relationship is created by adding Dynamic Properties: every added Property defines a bounding, where the name of the Property is the given route and the value of the Property is the
-    bounded Relationship.
-</p>
-
-<p>
-    Relationships are defined in a dynamic manner, based on the Property values. It is not allowed to bound routes to the already existing Relationships. These are: original, failure, unmatched.
-    It is allowed to transfer multiple routes to the same Relationship. In this case, the routes will share one outgoing FlowFile containing all the Records are assigned to the routes bounded to
-    the same Relationship. Records with partition not bounded to a Relationship are transferred to the "unmatched" Relationship within a single FlowFile.
-</p>
-
-<h3>Example</h3>
-
-<p>
-    This example shows a simple use case using ScriptedRouteRecord. The intended behaviour is to transfer records with "value" bigger than 1000 and equal or less than 1000 into a different
-    Relationship. In order to achieve this the first step is to provide a script which is capable to differentiate between the Records based on this condition. The following Groovy script will return
-    with "route1" for the first and "route2" with the second type of record.
-</p>
-
-<code>
-    if (record.getValue("value") > 1000) {
-    return "route1";
-    } else {
-    return "route2";
-    }
-</code>
-
-<p>
-    The ScriptedRouteRecord uses these returns values as routes. Records from the incoming FlowFile will be assigned into batches based on the partition returned by the script executed with the
-    given Record as argument. Consider the following Dynamic Properties were added to the processor:
-</p>
-
-<table>
-    <tr>
-        <th>Property Name</th>
-        <th>Property Value</th>
-    </tr>
-    <tr>
-        <td>route1</td>
-        <td>bigger</td>
-    </tr>
-    <tr>
-        <td>route2</td>
-        <td>smaller</td>
-    </tr>
-</table>
-
-<p>
-    Based on this, all the Records assigned to the "route1" route are transferred to the Relationship named "bigger" in one FlowFile. Accordingly all the ones assigned to the "route2"
-    are transferred to the "smaller" Relationship in the same fashion.
-</p>
-
-</body>
-</html>
